TST: Consolidated tests for the equation system
